# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de prueba NITSNETS",
        "description": "Esta es una primera versión de la API",
        "contact": "jorgepastor9120@gmail.com",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000"
        }
    ],
    "paths": {
...
        "/api/v1/members": {
            "get": {
                "tags": [
                    "Members"
                ],
                "summary": "Mostrar el listado de socios",
                "description": "Mostramos el listado de los socios solicitados.",
                "operationId": "f123b24777b50b0ba6fd1941011247bc",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "No se ha podido cargar el listado."
                    },
                    "503": {
                        "description": "El servidor no está disponible en este momento"
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Members"
                ],
                "summary": "Registra un socio",
                "description": "Registrar un socio.",
                "operationId": "390c26a692d9bcf656f35603e9da4dee",
                "requestBody": {
                    "description": "Registra un socio",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "Prueba"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "prueba@gmail.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "201": {
                        "description": "El socio ha sido creado"
                    },
                    "404": {
                        "description": "No se ha podido registrar el socio."
                    },
                    "503": {
                        "description": "El servidor no está disponible en este momento"
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/api/v1/members/{id}": {
            "put": {
                "tags": [
                    "Members"
                ],
                "summary": "Actualiza un socio",
                "description": "Actualizar un registro.",
                "operationId": "ac90f5a9c6d6dd9c699389c72facebee",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id member",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int20"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Actualiza un socio",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "Prueba"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "prueba@gmail.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "No se ha podido actualizar el registro."
                    },
                    "405": {
                        "description": "No se ha permitido actualizar el registro."
                    },
                    "503": {
                        "description": "El servidor no está disponible en este momento"
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Members"
                ],
                "summary": "Elimina un socio",
                "description": "Elimina un registro.",
                "operationId": "d17879bfe9b223bba4102503b362dc7d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id member",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int20"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Elimina un socio"
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "No se ha podido eliminar el registro."
                    },
                    "503": {
                        "description": "El servidor no está disponible en este momento"
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "passport": {
                "type": "oauth2",
                "description": "Laravel passport oauth2 security.",
                "in": "header",
                "scheme": "http",
                "name": "authorization",
                "flows": {
                    "password": {
                        "authorizationUrl": "http://localhost/oauth/authorize",
                        "tokenUrl": "http://127.0.0.1:8000/api/v1/oauth/token",
                        "refreshUrl": "http://localhost/token/refresh",
                        "scopes": []
                    }
                }
            }
        }
    },
...
}